Okay, finally figured it out and fixed 9B.
It's weird (and frustrating) that this didn't come up until
now.  I made a simplifying assumption that worked in part A,
but it took me forever to realize it in part B.  I didn't move
a knot correctly when the knot in front of it moved diagonally.
Now I do.  My solution isn't the prettiest, but it works.